Stick with node-tap 16 for now
Forced coverage metrics and 100% coverage without any way to override is
stupid, so we are going to stick with node-tap 16 until version 18 is
fixed.

If it does not get fixed, we have think about making our own tap runner
with assertions. Something closer to Test::More would be nice anyway,
and there's a lot of room for improvements in the API.